requirements.txt
kivy
plyer
anthropic
python-dotenv